Nav bar and proper elements selectable
