+++ src/components/nav/NavBar.js
@@ -1,0 +1,29 @@
+import React from "react"
+import { Link } from "react-router-dom"
+import "./NavBar.css"
+
+// Create the NavBar component for the constant navigation elements
+export const NavBar = (props) => {
+    return (
+        // Notice the use of the <Link/> component. This comes from the React Router package you installed. 
+        // It has an attribute named to. It will render a hyperlink in your DOM, and when clicked, 
+        // it will change the URL in the browser to the value of the to attribute.
+        <ul className="navbar">
+            <li className="navbar__item active">
+                <Link className="navbar__link" to="/">NSS Kennels</Link>
+            </li>
+            <li className="navbar__item">
+                <Link className="navbar__link" to="/locations">Locations</Link>
+            </li>
+            <li className="navbar__item">
+                <Link className="navbar__link" to="/animals">Animals</Link>
+            </li>
+            <li className="navbar__item">
+                <Link className="navbar__link" to="/customers">Customers</Link>
+            </li>
+            <li className="navbar__item">
+                <Link className="navbar__link" to="/employees">Employees</Link>
+            </li>
+        </ul>
+    )
+}
