+++ src/index.js
@@ -1,12 +1,18 @@
 import React from "react"
 import ReactDOM from "react-dom"
 import { Kennel } from "./components/Kennel"
+import { BrowserRouter as Router } from "react-router-dom"
+import "./index.css"
 import reportWebVitals from './reportWebVitals';
 
+// With Router code in place, you should be able to navigate between multiple views and display different HTML representations.
 ReactDOM.render(
   <React.StrictMode>
-    <Kennel />
+    <Router>
+      <Kennel />
+    </Router>
   </React.StrictMode>,
+  
   document.getElementById('root')
 );
